requirements.txt
absl-py==1.3.0
amqp==5.1.1
asgiref==3.6.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
azure-common==1.1.28
azure-core==1.26.3
azure-mgmt-core==1.3.2
azure-mgmt-notificationhubs==8.0.0
azure-servicebus==7.8.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
boto3==1.26.13
botocore==1.29.13
bs4==0.0.1
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
contourpy==1.0.5
cryptography==39.0.1
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
Django==4.1.4
django-allauth==0.52.0
django-cors-headers==3.13.0
django-crontab==0.7.1
django-import-export==3.0.2
django-phonenumber-field==7.0.1
django-picklefield==3.1
django-storages==1.13.1
django-timezone-field==5.0
django-upgrade==1.10.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.0.1
filelock==3.9.0
flatbuffers==22.9.24
fonttools==4.37.4
gast==0.4.0
gevent==22.10.2
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-images-download==2.8.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.49.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.7.0
html5lib==1.1
idna==3.4
importlib-metadata==5.0.0
isodate==0.6.1
jmespath==1.0.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kombu==5.2.4
libclang==14.0.6
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.6.1
MouseInfo==0.1.3
msrest==0.7.1
Naked==0.1.32
numpy==1.23.4
oauthlib==3.2.2
odfpy==1.4.1
openai==0.27.2
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.3
phonenumbers==8.13.3
Pillow==9.2.0
pip-upgrade==0.0.6
platformdirs==3.1.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.17
PyGetWindow==0.0.9
PyJWT==2.6.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytweening==1.0.4
pytz==2022.7
#pywin32==305
PyYAML==6.0
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
selenium==4.6.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.3.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
# tensorflow==2.10.0
# tensorflow-estimator==2.10.0
# tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tk==0.1.0
tokenize-rt==5.0.0
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.5.0
tzdata==2022.7
uamqp==1.6.4
urllib3==1.26.14
vine==5.0.0
virtualenv==20.20.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.2
whitenoise==6.3.0
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.9.0
zope.event==4.6
zope.interface==5.5.2
